+++ product-sdk/packages/tx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @parity/product-sdk-tx
 
+## 0.2.12
+
+### Patch Changes
+
+- @parity/product-sdk-keys@0.3.8
+
 ## 0.2.11
 
 ### Patch Changes